add test for floorplan import
