+++ test/e2e/fixtures/floorplan-2017-08-31.json
@@ -1,0 +1,298 @@
+{
+  "application": {
+    "currentSelections": {
+      "building_unit_id": null,
+      "mode": "spaces",
+      "snapMode": "grid-strict",
+      "space_type_id": null,
+      "story": null,
+      "story_id": "1",
+      "subselection_id": "13",
+      "thermal_zone_id": null,
+      "tool": "Rectangle"
+    },
+    "modes": [
+      "spaces",
+      "shading",
+      "building_units",
+      "thermal_zones",
+      "space_types",
+      "images"
+    ],
+    "scale": {
+      "x": {
+        "pixels": 1531,
+        "rwuRange": [
+          -250,
+          250
+        ]
+      },
+      "y": {
+        "pixels": 534,
+        "rwuRange": [
+          -87.19790986283475,
+          87.19790986283475
+        ]
+      }
+    },
+    "tools": [
+      "Pan",
+      "Drag",
+      "Rectangle",
+      "Polygon",
+      "Eraser",
+      "Select",
+      "Map",
+      "Fill"
+    ]
+  },
+  "building_units": [],
+  "construction_sets": [],
+  "daylighting_controls": [],
+  "project": {
+    "config": {
+      "language": "EN-US",
+      "north_axis": 0,
+      "units": "ft"
+    },
+    "grid": {
+      "spacing": 4,
+      "visible": true
+    },
+    "map": {
+      "elevation": 0,
+      "enabled": true,
+      "initialized": true,
+      "latitude": 39.765300000000025,
+      "longitude": -104.98630000000001,
+      "rotation": 0,
+      "visible": true,
+      "zoom": 4.5
+    },
+    "previous_story": {
+      "visible": true
+    },
+    "showImportExport": false,
+    "transform": {
+      "k": 1,
+      "x": 0,
+      "y": 0
+    },
+    "view": {
+      "max_x": 250,
+      "max_y": 87.19790986283475,
+      "min_x": -250,
+      "min_y": -87.19790986283475
+    }
+  },
+  "space_types": [],
+  "stories": [
+    {
+      "above_floor_plenum_height": 0,
+      "below_floor_plenum_height": 0,
+      "color": "",
+      "floor_to_ceiling_height": 0,
+      "geometry": {
+        "edges": [
+          {
+            "face_ids": [
+              "12"
+            ],
+            "id": "8",
+            "vertex_ids": [
+              "4",
+              "5"
+            ]
+          },
+          {
+            "face_ids": [
+              "12"
+            ],
+            "id": "9",
+            "vertex_ids": [
+              "5",
+              "6"
+            ]
+          },
+          {
+            "face_ids": [
+              "12"
+            ],
+            "id": "10",
+            "vertex_ids": [
+              "6",
+              "7"
+            ]
+          },
+          {
+            "face_ids": [
+              "12"
+            ],
+            "id": "11",
+            "vertex_ids": [
+              "7",
+              "4"
+            ]
+          },
+          {
+            "face_ids": [
+              "22"
+            ],
+            "id": "18",
+            "vertex_ids": [
+              "14",
+              "15"
+            ]
+          },
+          {
+            "face_ids": [
+              "22"
+            ],
+            "id": "19",
+            "vertex_ids": [
+              "15",
+              "16"
+            ]
+          },
+          {
+            "face_ids": [
+              "22"
+            ],
+            "id": "20",
+            "vertex_ids": [
+              "16",
+              "17"
+            ]
+          },
+          {
+            "face_ids": [
+              "22"
+            ],
+            "id": "21",
+            "vertex_ids": [
+              "17",
+              "14"
+            ]
+          }
+        ],
+        "faces": [
+          {
+            "edge_ids": [
+              "8",
+              "9",
+              "10",
+              "11"
+            ],
+            "edge_order": [
+              1,
+              1,
+              1,
+              1
+            ],
+            "id": "12"
+          }
+        ],
+        "id": "2",
+        "vertices": [
+          {
+            "edge_ids": [
+              "8",
+              "11"
+            ],
+            "id": "4",
+            "x": -108,
+            "y": 24
+          },
+          {
+            "edge_ids": [
+              "8",
+              "9"
+            ],
+            "id": "5",
+            "x": 16,
+            "y": 24
+          },
+          {
+            "edge_ids": [
+              "9",
+              "10"
+            ],
+            "id": "6",
+            "x": 16,
+            "y": -28
+          },
+          {
+            "edge_ids": [
+              "10",
+              "11"
+            ],
+            "id": "7",
+            "x": -108,
+            "y": -28
+          },
+          {
+            "edge_ids": [
+              "18",
+              "21"
+            ],
+            "id": "14",
+            "x": -172,
+            "y": 48
+          },
+          {
+            "edge_ids": [
+              "18",
+              "19"
+            ],
+            "id": "15",
+            "x": -128,
+            "y": 48
+          },
+          {
+            "edge_ids": [
+              "19",
+              "20"
+            ],
+            "id": "16",
+            "x": -128,
+            "y": -8
+          },
+          {
+            "edge_ids": [
+              "20",
+              "21"
+            ],
+            "id": "17",
+            "x": -172,
+            "y": -8
+          }
+        ]
+      },
+      "handle": "{1f360615-4031-4838-a1b4-b13e2fb6930c}",
+      "id": "1",
+      "images": [],
+      "multiplier": 0,
+      "name": "Story 1",
+      "shading": [],
+      "spaces": [
+        {
+          "building_unit_id": null,
+          "color": "#007373",
+          "construction_set_id": null,
+          "daylighting_controls": [],
+          "face_id": "12",
+          "handle": "{9c950a34-a709-4e7b-b1aa-55ec90226d68}",
+          "id": "3",
+          "name": "Space 1 - 1",
+          "space_type_id": null,
+          "thermal_zone_id": null,
+          "type": "space"
+        }
+      ],
+      "windows": []
+    }
+  ],
+  "thermal_zones": [],
+  "version": "0.3.0",
+  "windows": []
+}
