+++ test/e2e/specs/test.js
@@ -2,27 +2,7 @@
 // http://nightwatchjs.org/guide#usage
 const d3 = require('d3');
 const _ = require('lodash');
-
-function failOnError(browser) {
-  let errorOccurred;
-  browser.setFlagOnError = () => {
-    browser.execute(
-      'window.errorOccurred = false; window.addEventListener("error", () => { window.errorOccurred = true; })',
-    );
-    return browser;
-  };
-  browser.checkForErrors = () => {
-    browser.execute('return window.errorOccurred', [], ({ value }) => {
-      errorOccurred = value;
-    })
-    .perform((client, done) => {
-      client.assert.ok(!errorOccurred);
-      done();
-    });
-    return browser;
-  };
-  return browser;
-}
+const failOnError = require('../helpers').failOnError;
 
 module.exports = {
   'make space and new story': (browser) => {
